refactor: generalise the way _base.strict is calculated.
* specs/_base_spec.yaml: Be a lot more careful about accessing
symbol values after running `_base.strict` and obliterating the
execution environment as a result.
* lib/std/normalize/_base.lua (_ENV): Capture global _DEBUG into
the module environment.
(strict): Always set the callers environment to the argument table
so that the caller gets consistent behaviour.
(_DEBUG): When `std.debug_init` is not available, set the `strict`
field of the module `_DEBUG` symbol with the same `choose` function
that stdlib itself uses.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>