init: support optional level argument to __call metamethod.
* lib/std/normalize/_strict.lua (strict): Follow the same API is
"std.strict" and do not set the caller's environment.
(__call): Likewise, but by supporting an optional level argument.
* lib/std/normalize/init.lua (__call): Similarly, provide an
optional level argument so that wrapper functions that set the
stack level correctly can be written.
* specs/init_spec.yaml: Now that the environment of the correct
stack frame is set on Lua 5.1 hosts, we have to follow the
semantics of normalize correctly and always assign the result
back to `_ENV`.  And because the current release of Specl mangles
the in-process environment naively, rewrite broken examples to run
in a Lua subprocess.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>