normalize: correctly respect presence of std.strict.
* spec/spec_helper.lua (have_std_strict): New variable.
* spec/normalize_spec.yaml: New spec with examples of undeclared
accesses being diagnosed, when have_std_strict is set.
* lib/std/normalize/_strict.lua (setfenv): Need to set this in module
global scope so that it is bound when called from module functions in
Lua 5.1 hosts after setting a function environment that doesn't have
a setfenv entry.
(__call): Remember to set a strict environment when requested.
* lib/std/normalize/init.lua (_ENV): Don't let environment entries
default to nil values, lest they be mistake for undeclared access in
strict mode.
* lib/std/normalize/_base.lua (_ENV): Likewise.
(tointeger): Adjust to accept a true boolean argument accordingly.
* spec/init_spec.yaml (loads core functions): Use pairs instead of
next to make sure the accesses are forwarded to the proxy table.
* NEWS.md (Bug fixes): Update.

Signed-off-by: Gary V. Vaughan <gary@gnu.org>